File Structure: Clean up file & folder structure
