+++ frontend/src/components/IssueSetPage.tsx
@@ -7,17 +7,17 @@ import Notification from 'grommet/components/Notification';
 import Distribution from 'grommet/components/Distribution';
 
 import IssuanceOrderForm from './IssuanceOrderForm';
-import { getSetProtocolInstance } from './setProtocol';
-import { setMap } from './data/sets';
-import { api } from './api';
+import { getSetProtocolInstance } from '../utils/setProtocol';
+import { setMap } from '../constants/sets';
+import { api } from '../utils/api';
 import { BigNumber } from '0x.js';
 
-export interface CreateSetPageState {
+export interface IssueSetPageState {
     isLoading: boolean;
     errorMessage: string;
 }
 
-class CreateSetPage extends React.Component<RouteProps, CreateSetPageState> {
+class IssueSetPage extends React.Component<RouteProps, IssueSetPageState> {
     constructor(props) {
         super(props);
         this.state = {
@@ -25,7 +25,6 @@ class CreateSetPage extends React.Component<RouteProps, CreateSetPageState> {
             errorMessage: '',
         };
     }
-
     public render(): React.ReactNode {
         const setId = this.props.match.params.setId;
         const set = setMap[setId];
@@ -106,4 +105,4 @@ class CreateSetPage extends React.Component<RouteProps, CreateSetPageState> {
     };
 }
 
-export default CreateSetPage;
+export default IssueSetPage;
