+++ frontend/src/components/Logo.tsx
