+++ frontend/src/utils/web3Wrapper.ts
@@ -1,5 +1,5 @@
 import { Web3Wrapper } from '@0xproject/web3-wrapper';
 
-import { getProvider } from './provider';
+import { getProvider } from './getProvider';
 
 export const web3Wrapper = new Web3Wrapper(getProvider());
